+++ frontend/src/api/queueEndpoints.js
@@ -1,0 +1,2 @@
+export const getQueue = api => async queueId =>
+  await api.get(`programs/${queueId}/queue/`)
